Migrate experimental.testing and delete the experimental module.
