--- numba_dpex/experimental/__init__.py
@@ -1,7 +1,0 @@
-# SPDX-FileCopyrightText: 2023 - 2024 Intel Corporation
-#
-# SPDX-License-Identifier: Apache-2.0
-
-"""Contains experimental features that are meant as engineering preview and not
-yet production ready.
-"""
